dpkg-split: Do not corrupt binary part metadata on 32-bit systems
On 32-bit systems the format string used to generate the metadata
archive member of a split package specifies a 32-bit value to be
retrieved from va_arg, but the variable passed is a 64-bit one, which
messes with the layout of the stack, producing garbage on output.

Regression introduced in commit f2115151c19ff37b305296c23225807b0832086b.

+++ dpkg-split/split.c
@@ -206,7 +206,7 @@ mksplit(const char *file_src, const char *prefix, size_t partsize,
 		/* Write the debian-split part. */
 		varbufprintf(&partmagic, "%s\n%s\n%s\n%s\n%zu\n%zu\n%d/%d\n",
 		             SPLITVERSION, package, version, hash,
-		             st.st_size, partsize, curpart, nparts);
+		             (size_t)st.st_size, partsize, curpart, nparts);
 		dpkg_ar_member_put_mem(file_dst.buf, fd_dst, PARTMAGIC,
 		                       partmagic.buf, partmagic.used);
 		varbufreset(&partmagic);
